+++ apps/docs/components/announcement.tsx
@@ -1,5 +1,3 @@
-import { Separator } from "@/registry/default/ui/separator";
-import { ArrowRightIcon } from "@radix-ui/react-icons";
 import Link from "next/link";
 
 export function Announcement() {
@@ -12,7 +10,6 @@ export function Announcement() {
       <span className="hidden sm:inline">
         We're currently building the docs
       </span>
-      {/* <ArrowRightIcon className="ml-1 h-4 w-4" /> */}
     </Link>
   );
 }