+++ apps/docs/registry/default/example/register-form-demo.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+import { RegisterForm } from "@sikka/hawa/blocks";
+
+export default function LoginFormDemo() {
+  return <RegisterForm onRegister={(data) => console.log("data is ", data)} />;
+}